docker-compose install existance check

+++ carnival_contrib/docker.py
@@ -32,17 +32,22 @@ class CeInstallUbuntu(Step):
 
 
 class ComposeInstall(Step):
-    def run(self, docker_compose_version="1.25.1", docker_compose_dest="/usr/local/bin/docker-compose"):
+    def run(self, docker_compose_version="1.25.1", docker_compose_dest="/usr/local/bin/docker-compose", force=False):
         """
         Install docker-compose
 
         :param docker_compose_version: compose version
         :param docker_compose_dest: install directory
         """
-        log(f"Installing compose...")
-        download_link = f"https://github.com/docker/compose/releases/download/{docker_compose_version}/docker-compose-`uname -s`-`uname -m`"
-        cmd.cli.run(f"sudo curl -sL {download_link} -o {docker_compose_dest}")
-        cmd.cli.run(f"sudo chmod a+x {docker_compose_dest}")
+        from carnival.cmd import fs
+
+        if not fs.is_file_exists(docker_compose_dest) or force:
+            log(f"Installing docker-compose...")
+            download_link = f"https://github.com/docker/compose/releases/download/{docker_compose_version}/docker-compose-`uname -s`-`uname -m`"
+            cmd.cli.run(f"sudo curl -sL {download_link} -o {docker_compose_dest}")
+            cmd.cli.run(f"sudo chmod a+x {docker_compose_dest}")
+        else:
+            log(f"docker-compose already installed...")
 
 
 class DockerUploadImageFile(Step):
